[Refactor] Model 구분 및 정리
