+++ Gongbu/Model/AttendenceInfoModel.swift
@@ -1,0 +1,14 @@
+//
+//  AttendenceInfoModel.swift
+//  Gongbu
+//
+//  Created by kimhyeongmin on 2022/06/12.
+//
+
+import Foundation
+
+struct AttendenceInfo {
+    var numberOfAttendance: Int
+    var numberOfAbsent: Int
+    var numberOfLate: Int
+}
